% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{bootstrap}
\alias{bootstrap}
\title{Generates the bootstrapped heirarchical clustering of the given LSI matrix using named method for clustering}
\usage{
bootstrap(lsi_matrix, method, nboot = 1000,
  distance_method = "euclidean")
}
\arguments{
\item{lsi_matrix}{The matrix to cluster. Generated from lsi_matrix()}

\item{method}{The name of the method to use. See the documentation for hclust() for options}

\item{nboot}{The number of bootstrap replications. The default is 1000.}

\item{distance_method}{the distance measure to be used. This should be a character string, or a function which returns a dist object. A character string should be (an abbreviation of) one of "correlation", "uncentered", "abscor" or those which are allowed for method argument in dist function. The default is "correlation". See details section in this help and method argument in dist}
}
\value{
???
}
\description{
Generates the bootstrapped heirarchical clustering of the given LSI matrix using named method for clustering
}
\examples{
bootstrap(lsi_matrix("data.txt"), "average")
bootstrap(lsi_matrix("data.txt"), "average", nboot=500, distance_method="uncentered")
}
