% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{cluster}
\alias{cluster}
\title{Produces a plot of the heirarchical clustering of the given LSI matrix using the named method for clustering}
\usage{
cluster(lsi_matrix, method)
}
\arguments{
\item{lsi_matrix}{The matrix to cluster. Generated from lsi_matrix()}

\item{method}{The name of the method to use. See the documentation for hclust() for options}
}
\description{
Produces a plot of the heirarchical clustering of the given LSI matrix using the named method for clustering
}
\examples{
cluster(lsi_matrix("data.txt"), "average")
}
