% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{compare_songs}
\alias{compare_songs}
\title{Compare Songs}
\usage{
compare_songs(filename, cost_matrix = NULL, fileEncoding = "",
  fleven = leven)
}
\arguments{
\item{filename}{The file containing the input data. See the description for lsi_matrix() for the format of the input data}

\item{cost_matrix}{An optional cost matrix to use when computing the Levenshtein distance}

\item{fileEncoding}{character string: if non-empty declares the encoding used on the file so the character data can be read correctly. See the 'Encoding' section of the help for file}

\item{fleven}{The Levenshtein-calculating function to use. Defaults to leven (slow). You can supply fleven as the argument to use the fast (native) code}
}
\value{
A list with values: \itemize{
\item lsi_matrix: the generated LSI matrix - see \link{lsi_matrix}
\item theme_matrix: A matrix where cell i,j is the average LSI between theme i and theme j
\item set_medians: A list of median distances for each theme. Each entry contains these attributes: fixme
\item median_lsi_matrix: The LSI matrix between medians for each theme. Cell i,j in this matrix is the LSI of the median of theme i and the median of theme j}
}
\description{
Compare all song strings in the given file
}
\examples{
compare_songs("data.txt", fileEncoding="utf8", fleven=fleven)
}
