% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{cost_matrix_from_file}
\alias{cost_matrix_from_file}
\title{Load a cost matrix}
\usage{
cost_matrix_from_file(filename, threshhold = 1, exponentialscale = 1)
}
\arguments{
\item{filename}{?}

\item{threshhold}{?}

\item{exponentialscale}{?}
}
\value{
The cost matrix
}
\description{
This function loads a cost matrix from a file.
}
\examples{
cost_matrix_from_file("matrix.txt")
}
