% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{fast_leven}
\alias{fast_leven}
\title{Fast Levenshtein calculation}
\usage{
fast_leven(x, y, cost_map = NULL)
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}

\item{cost_map}{An optional cost map specifying the cost of specific distributions. Not to be confused with a cost matrix.}
}
\value{
The distance between x and y
}
\description{
This function is optimised to compute the Levenshtein distance as quickly as possible
}
\examples{
leven(c("k", "i", "t", "t", "e", "n"), c("s", "i", "t", "t", "i", "n", "g"))
}
\seealso{
\link{cost_matrix_from_file}

\link{optimise_cost_matrix}
}
