% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{optimise_cost_matrix}
\alias{optimise_cost_matrix}
\title{Generate a cost map}
\usage{
optimise_cost_matrix(cost_matrix)
}
\arguments{
\item{cost_matrix}{The cost matrix to convert}
}
\value{
The generated cost map
}
\description{
Product a cost map from a cost matrix. This function is required to product a fast-lookup cost map for supply to fast_leven
}
\examples{
optimise_cost_matrix(cost_matrix_from_file("matrix.txt"))
}
\seealso{
\link{cost_matrix_from_file}
}
