% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{fast_leven}
\alias{fast_leven}
\title{Compute the Levenshtein distance between two vectors (quickly)}
\usage{
fast_leven(x, y, cost_matrix = NULL)
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}

\item{cost_matrix}{An optional cost matrix to specify the cost of specific substitutions.}
}
\value{
The distance between x and y
}
\description{
This function is optimised to run as quickly as possible
}
\examples{
leven(c("k", "i", "t", "t", "e", "n"), c("s", "i", "t", "t", "i", "n", "g"))
}
